+++ forge/src/main/java/top/theillusivec4/polymorph/common/integration/fastfurnace/FastFurnaceModule.java
@@ -23,12 +23,11 @@ package top.theillusivec4.polymorph.common.integration.fastfurnace;
 
 import net.minecraft.item.crafting.AbstractCookingRecipe;
 import net.minecraft.item.crafting.IRecipe;
+import net.minecraft.tileentity.AbstractFurnaceTileEntity;
 import net.minecraft.tileentity.TileEntity;
-import shadows.fastfurnace.tile.TileFastBlastFurnace;
-import shadows.fastfurnace.tile.TileFastFurnace;
-import shadows.fastfurnace.tile.TileFastSmoker;
+import org.apache.commons.lang3.reflect.FieldUtils;
+import top.theillusivec4.polymorph.common.PolymorphMod;
 import top.theillusivec4.polymorph.common.integration.AbstractCompatibilityModule;
-import top.theillusivec4.polymorph.mixin.integration.fastfurnace.AccessorTileFastFurnace;
 
 public class FastFurnaceModule extends AbstractCompatibilityModule {
 
@@ -36,9 +35,16 @@ public class FastFurnaceModule extends AbstractCompatibilityModule {
   public boolean selectRecipe(TileEntity tileEntity, IRecipe<?> recipe) {
 
     if (recipe instanceof AbstractCookingRecipe &&
-        (tileEntity instanceof TileFastFurnace || tileEntity instanceof TileFastSmoker ||
-            tileEntity instanceof TileFastBlastFurnace)) {
-      ((AccessorTileFastFurnace) tileEntity).setCurRecipe((AbstractCookingRecipe) recipe);
+        tileEntity instanceof AbstractFurnaceTileEntity) {
+      try {
+        FieldUtils.writeField(tileEntity, "curRecipe", recipe, true);
+      } catch (IllegalAccessException e) {
+        PolymorphMod.LOGGER.error("Cannot access curRecipe in FastFurnace!");
+        e.printStackTrace();
+      } catch (IllegalArgumentException e) {
+        PolymorphMod.LOGGER.error("Cannot find curRecipe in FastFurnace!");
+        e.printStackTrace();
+      }
     }
     return false;
   }